[REL] stock_ean128 to V11. [FIX] Rename py and xml files. [FIX] Separate classes and views in respective files. [FIX] Update README.
