+++ stock_ean128/__manifest__.py
@@ -19,7 +19,7 @@
 ##############################################################################
 {
     'name': 'Stock EAN128',
-    'version': '9.0.1.0.0',
+    'version': '11.0.1.0.0',
     'category': 'Warehouse Management',
     'sequence': 14,
     'summary': '',
@@ -29,15 +29,15 @@
     'images': [
     ],
     'depends': [
-        'product_expiry',
+        'stock',
     ],
     'data': [
-        'views/stock_view.xml',
-        'views/product_view.xml',
+        'views/stock_production_lot.xml',
+        'views/product_template.xml',
     ],
     'demo': [
     ],
-    'installable': False,
+    'installable': True,
     'auto_install': False,
     'application': False,
 }
