+++ stock_ean128/models/__init__.py
@@ -3,5 +3,6 @@
 # directory
 ##############################################################################
 
-from . import product
-from . import stock
+from . import product_product
+from . import product_template
+from . import stock_production_lot
