+++ stock_ean128/models/stock_production_lot.py
@@ -8,22 +8,25 @@ from odoo import models, fields, api
 class StockProductionLot(models.Model):
     _inherit = 'stock.production.lot'
 
-    @api.one
-    @api.depends('name', 'product_id', 'product_id.default_code')
-    def action_compute(self):
-        name = ''
-        if self.product_id.default_code:
-            name += ' 01 ' + self.product_id.default_code
-        name += ' 10 ' + self.name
-        # if self.life_date:
-        #     life_date = fields.Datetime.from_string(self.life_date)
-        #     name += ' 17 ' + life_date.strftime('%d%m%y')
-        # else:
-        #     name += ' 17 ' + 'N.A'
-        self.ean_128 = name
-
     ean_128 = fields.Char(
-        string="EAN128", compute='action_compute', store=True)
+        string="EAN128",
+        compute='_compute_action_compute',
+        store=True,
+    )
+
+    @api.depends('name', 'product_id', 'product_id.default_code')
+    def _compute_action_compute(self):
+        for rec in self:
+            name = ''
+            if rec.product_id.default_code:
+                name += ' 01 ' + rec.product_id.default_code
+            name += ' 10 ' + rec.name
+            # if self.life_date:
+            #     life_date = fields.Datetime.from_string(self.life_date)
+            #     name += ' 17 ' + life_date.strftime('%d%m%y')
+            # else:
+            #     name += ' 17 ' + 'N.A'
+            rec.ean_128 = name
 
     @api.model
     def name_search(
