+++ stock_ean128/views/product_template.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<odoo>
+    <record id="view_product_template_search" model="ir.ui.view">
+        <field name="name">product search</field>
+        <field name="model">product.template</field>
+        <field name="inherit_id" ref="product.product_template_search_view"/>
+        <field name="arch" type="xml">
+            <field name="name" position="after">
+                <field name="lot_ids"/>
+            </field>
+        </field>
+    </record>
+</odoo>
