--- stock_ean128/views/product_view.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <data>
-
-        <record id="view_product_template_search" model="ir.ui.view">
-            <field name="name">product search</field>
-            <field name="model">product.template</field>
-            <field name="inherit_id" ref="product.product_template_search_view"/>
-            <field name="arch" type="xml">
-               <field name="name" position="after">
-                    <field name="lot_ids"/>
-                </field>
-            </field>  
-        </record>
-
-    </data>
-</odoo>
