+++ stock_ean128/views/stock_production_lot.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0"?>
+<odoo>
+    <record id="view_stock_lot_search" model="ir.ui.view">
+        <field name="name">Stock production lot</field>
+        <field name="model">stock.production.lot</field>
+        <field name="inherit_id" ref="stock.search_product_lot_filter"/>
+        <field name="arch" type="xml">
+            <field name="product_id" position="before">
+                <field name="ean_128"/>
+            </field>
+        </field>
+    </record>
+
+    <record id="view_stock_lot" model="ir.ui.view">
+        <field name="name">Stock production lot</field>
+        <field name="model">stock.production.lot</field>
+        <field name="inherit_id" ref="stock.view_production_lot_form"/>
+        <field name="arch" type="xml">
+            <field name="ref" position="after">
+                <field name="ean_128"/>
+            </field>
+        </field>
+    </record>
+</odoo>
